backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
pydantic>=2.11
pydantic-settings>=2.8
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt==4.0.1
python-multipart>=0.0.20
email-validator>=2.2
python-dotenv>=1.0
openai>=1.50
anthropic>=0.52
python-docx>=1.1
pymupdf>=1.25
# DMAhub data-ops deps
pandas>=2.0
openpyxl>=3.1
numpy>=1.24
scikit-learn>=1.3
requests>=2.31
beautifulsoup4>=4.12
nltk>=3.8
cryptography>=42.0
